fix: only call /users for google social login
