+++ cmd/terramate/cli/cloud.go
@@ -4,12 +4,10 @@
 package cli
 
 import (
-	"context"
 	"net/http"
 	"time"
 
 	"github.com/golang-jwt/jwt"
-	"github.com/terramate-io/terramate/cloud"
 	"github.com/terramate-io/terramate/cmd/terramate/cli/cliconfig"
 	"github.com/terramate-io/terramate/cmd/terramate/cli/out"
 	"github.com/terramate-io/terramate/errors"
@@ -29,10 +27,9 @@ type credential interface {
 	Token() (string, error)
 	Refresh() error
 	Claims() jwt.MapClaims
-	DisplayClaims() []keyValue
 	IsExpired() bool
 	ExpireAt() time.Time
-	String() string
+	Info(cloudcfg cloudConfig) error
 }
 
 type keyValue struct {
@@ -54,16 +51,18 @@ func (c *cli) cloudInfo() {
 	}
 
 	cloud := cloudConfig{
-		baseAPI:    cloudBaseURL,
-		credential: cred,
-		client:     &http.Client{},
-		output:     c.output,
+		baseAPI: cloudBaseURL,
+		client:  &http.Client{},
+		output:  c.output,
 	}
 
-	err = cloud.Info()
+	err = cred.Info(cloud)
 	if err != nil {
-		c.output.MsgStdErr("error: %v", err)
+		fatal(err)
 	}
+
+	// verbose info
+	cloud.output.MsgStdOutV("next token refresh in: %s", time.Until(cred.ExpireAt()))
 }
 
 func (c *cli) loadCredential() (credential, error) {
@@ -99,74 +98,4 @@ func tokenClaims(token string) (jwt.MapClaims, error) {
 		return claims, nil
 	}
 	return nil, errors.E("invalid jwt token claims")
-}
-
-func (cloudcfg *cloudConfig) Info() error {
-	cred := cloudcfg.credential
-	client := cloud.Client{
-		BaseURL:    cloudBaseURL,
-		Credential: cred,
-	}
-
-	const apiTimeout = 5 * time.Second
-
-	var (
-		err  error
-		user cloud.User
-		orgs cloud.MemberOrganizations
-	)
-
-	func() {
-		ctx, cancel := context.WithTimeout(context.Background(), apiTimeout)
-		defer cancel()
-		orgs, err = client.MemberOrganizations(ctx)
-	}()
-
-	if err != nil {
-		return err
-	}
-
-	func() {
-		ctx, cancel := context.WithTimeout(context.Background(), apiTimeout)
-		defer cancel()
-		user, err = client.Users(ctx)
-	}()
-
-	if err != nil && !errors.IsKind(err, cloud.ErrNotFound) {
-		return err
-	}
-
-	userErr := err
-
-	if len(orgs) == 0 && cred.Name() == githubOIDCProviderName {
-		cloudcfg.output.MsgStdOut("status: untrusted")
-	} else {
-		cloudcfg.output.MsgStdOut("status: signed in")
-	}
-
-	cloudcfg.output.MsgStdOut("provider: %s", cred.Name())
-
-	if userErr == nil && user.DisplayName != "" {
-		cloudcfg.output.MsgStdOut("user: %s", user.DisplayName)
-	}
-
-	for _, kv := range cred.DisplayClaims() {
-		cloudcfg.output.MsgStdOut("%s: %s", kv.key, kv.value)
-	}
-
-	if len(orgs) > 0 {
-		cloudcfg.output.MsgStdOut("organizations: %s", orgs)
-	}
-
-	if user.DisplayName == "" {
-		cloudcfg.output.MsgStdErr("Warning: On-boarding is incomplete.  Please visit cloud.terramte.io to complete on-boarding.")
-	}
-
-	if len(orgs) == 0 {
-		cloudcfg.output.MsgStdErr("Warning: You are not part of an organization. Please visit cloud.terramate.io to create an organization.")
-	}
-
-	// verbose info
-	cloudcfg.output.MsgStdOutV("next token refresh in: %s", time.Until(cred.ExpireAt()))
-	return nil
 }
